change ksapi write_status status_code from int to uint16_t

+++ include/KPushGate.h
@@ -13,7 +13,7 @@ kgl_forward_stream*new_forward_stream(kgl_output_stream*down_gate);
 kgl_output_stream *get_check_output_stream();
 
 
-void forward_write_status(kgl_output_stream* gate, KREQUEST r, int status_code);
+void forward_write_status(kgl_output_stream* gate, KREQUEST r, uint16_t status_code);
 KGL_RESULT forward_write_body(kgl_output_stream* gate, KREQUEST r, const char* buf, int len);
 KGL_RESULT forward_write_header_finish(kgl_output_stream* gate, KREQUEST r);
 KGL_RESULT forward_write_message(kgl_output_stream* gate, KREQUEST rq, KGL_MSG_TYPE type, const void* msg, int32_t msg_flag);
